requirements.txt
# Core LangChain
langchain>=0.2.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0

# Integrations
langchain-groq>=1.0.0
langchain-chroma>=0.2.0
langchain-huggingface>=1.0.0

# Vector DB + local embeddings
chromadb>=0.5.0
sentence-transformers>=2.6.0

# PDF loader
pypdf>=4.0.0

# .env support
python-dotenv>=1.0.0
